update(types): add event types
