+++ package.json
@@ -7,7 +7,7 @@
     "example": "examples",
     "test": "test"
   },
-  "types": "lib/index.d.ts",
+  "types": "types/index.d.ts",
   "scripts": {
     "test": "standard && tap test/*.test.js"
   },
